tienda-pachito-mysql/Backend/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.4.0
mysqlclient==2.2.4
